requirements.txt
streamlit>=1.39
pandas>=2.2
plotly>=5.24
altair>=5.3
supabase>=2.4
python-dotenv>=1.0
pyarrow>=17.0
openpyxl>=3.1

